# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Karijera.rs API
  description: >
    REST API za platformu za oglašavanje i prijavu na poslove. Omogućava registraciju
    i autentikaciju korisnika (tražioci posla, kompanije, administratori), upravljanje
    profilima, objavljivanje i pretragu oglasa za posao, prijavljivanje na oglase i
    administraciju sistema.
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api
    description: Lokalni razvojni server
...
paths:
...
  /change-password:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Izmena lozinke
      description: Menja lozinku ulogovanog korisnika nakon provere trenutne lozinke.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                  format: password
                  example: "staraLozinka123"
                new_password:
                  type: string
                  format: password
                  minLength: 8
                  example: "novaLozinka123"
      responses:
        '200':
          description: Lozinka uspešno promenjena.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Trenutna lozinka nije ispravna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validacija nije prošla.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  # ===================== JOB LISTINGS (public) =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: >
        Pristupni token dobijen prilikom registracije ili prijave. Šalje se u
        zaglavlju "Authorization: Bearer {token}" za sve zaštićene rute.
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Validation failed."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...